+++ qase-cypress/src/metadata/models.ts
@@ -6,4 +6,20 @@ export interface Metadata {
   ignore?: boolean;
   suite?: string | undefined;
   comment?: string | undefined;
+  steps?: (StepStart | StepEnd)[];
+  currentStepId?: string | undefined;
+  firstStepName?: string | undefined;
+}
+
+export interface StepStart {
+  id: string;
+  timestamp: number;
+  name: string;
+  parentId: string | undefined;
+}
+
+export interface StepEnd {
+  id: string;
+  timestamp: number;
+  status: string;
 }